+++ app/src/main/java/com/example/catslist/database/CatDatabaseEntity.kt
@@ -1,4 +1,4 @@
-package com.example.catslist.room
+package com.example.catslist.database
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
