+++ app/src/main/java/com/example/catslist/database/CatsDao.kt
@@ -1,4 +1,4 @@
-package com.example.catslist.room
+package com.example.catslist.database
 
 import androidx.room.*
