+++ app/src/main/java/com/example/catslist/viewmodels/CatsListFragmentViewModel.kt
@@ -12,8 +12,8 @@ import androidx.appcompat.widget.AppCompatImageButton
 import androidx.lifecycle.AndroidViewModel
 import com.example.catslist.R
 import com.example.catslist.models.Cat
-import com.example.catslist.room.CatDatabaseEntity
-import com.example.catslist.room.CatsDatabaseRepository
+import com.example.catslist.database.CatDatabaseEntity
+import com.example.catslist.database.CatsDatabaseRepository
 import com.example.catslist.tools.CatStorage
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers.Main
